requirements.txt
cfscrape==2.1.1
google_api_python_client==2.48.0
google_auth_oauthlib==0.5.1
Pillow==9.1.1
protobuf==3.20.1
psutil==5.9.0
psycopg2==2.9.3
psycopg2_binary==2.9.3
Pyrogram==2.0.24
python-dotenv==0.20.0
python_magic==0.4.25
python_telegram_bot==13.11
requests==2.25.1
telegram==0.0.1
telegraph==2.1.0
TgCrypto
tenacity==8.0.1
urllib3==1.26.5
wget==3.2
yt_dlp==2022.5.18
